refactor: merge in dns-websys into dns crate
