+++ Cargo.toml
@@ -54,7 +54,6 @@ members = [
     "swarm-test",
     "swarm",
     "transports/dns",
-    "transports/dns-websys",
     "transports/noise",
     "transports/plaintext",
     "transports/pnet",
@@ -83,7 +82,6 @@ libp2p-connection-limits = { version = "0.7.0", path = "misc/connection-limits" 
 libp2p-core = { version = "0.44.0", path = "core" }
 libp2p-dcutr = { version = "0.15.0", path = "protocols/dcutr" }
 libp2p-dns = { version = "0.45.0", path = "transports/dns" }
-libp2p-dns-websys = { version = "0.1.0-alpha", path = "transports/dns-websys" }
 libp2p-floodsub = { version = "0.48.0", path = "protocols/floodsub" }
 libp2p-gossipsub = { version = "0.50.0", path = "protocols/gossipsub" }
 libp2p-identify = { version = "0.48.0", path = "protocols/identify" }
