--- transports/dns-websys/Cargo.toml
@@ -1,32 +1,0 @@
-[package]
-name = "libp2p-dns-websys"
-edition.workspace = true
-rust-version = { workspace = true }
-description = "DNS transport implementation via DNS-over-HTTPS for libp2p under WASM environment"
-version = "0.1.0-alpha"
-license = "MIT"
-repository = "https://github.com/libp2p/rust-libp2p"
-keywords = ["peer-to-peer", "libp2p", "networking"]
-categories = ["network-programming", "asynchronous"]
-
-[dependencies]
-futures = { workspace = true }
-js-sys = "0.3.77"
-libp2p-core = { workspace = true }
-parking_lot = "0.12.5"
-send_wrapper = { version = "0.6.0", features = ["futures"] }
-serde = { version = "1.0.228", features = ["derive"] }
-serde_json = "1.0.150"
-smallvec = "1.15.1"
-thiserror = { workspace = true }
-tracing = { workspace = true }
-url = "2.5.4"
-wasm-bindgen = "0.2.100"
-wasm-bindgen-futures = { workspace = true }
-web-sys = { version = "0.3.77", features = ["AbortSignal", "Headers", "Request", "RequestInit", "Response", "Window", "WorkerGlobalScope"] }
-
-[package.metadata.docs.rs]
-all-features = true
-
-[lints]
-workspace = true
