+++ transports/dns/Cargo.toml
@@ -13,13 +13,29 @@ categories = ["network-programming", "asynchronous"]
 [dependencies]
 futures = { workspace = true }
 libp2p-core = { workspace = true }
-libp2p-identity = { workspace = true }
 parking_lot = "0.12.5"
-hickory-resolver = { workspace = true, features = ["system-config"] }
 smallvec = "1.15.1"
 tracing = { workspace = true }
 
-[dev-dependencies]
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+hickory-resolver = { workspace = true, features = ["system-config"] }
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+getrandom = { workspace = true }
+js-sys = "0.3.77"
+send_wrapper = { version = "0.6.0", features = ["futures"] }
+serde = { version = "1.0.228", features = ["derive"] }
+serde_json = "1.0.150"
+thiserror = { workspace = true }
+url = "2.5.4"
+wasm-bindgen = "0.2.100"
+wasm-bindgen-futures = { workspace = true }
+web-sys = { version = "0.3.77", features = ["AbortSignal", "Headers", "Request", "RequestInit", "Response", "Window", "WorkerGlobalScope"] }
+
+[target.'cfg(target_arch = "wasm32")'.dev-dependencies]
+wasm-bindgen-test = "0.3.50"
+
+[target.'cfg(not(target_arch = "wasm32"))'.dev-dependencies]
 libp2p-identity = { workspace = true, features = ["rand"] }
 tokio = { workspace = true, features = ["rt", "time"] }
 tracing-subscriber = { workspace = true, features = ["env-filter"] }
@@ -31,6 +47,7 @@ tokio = ["hickory-resolver/tokio"]
 # More information: https://docs.rs/about/builds#cross-compiling
 [package.metadata.docs.rs]
 all-features = true
+targets = ["x86_64-unknown-linux-gnu", "wasm32-unknown-unknown"]
 
 [lints]
 workspace = true
